// src/app/api/dto/AccessLevel.g.ts
export enum AccessLevel {
    EmployeePageView = "EmployeePageView",
    EmployeeCreate = "EmployeeCreate",
    EmployeeEdit = "EmployeeEdit",
    EmployeeBlocking = "EmployeeBlocking",
    RolesPageView = "RolesPageView",
    RolesRolesManage = "RolesRolesManage",
    CardPageView = "CardPageView",
    CardInfoEdit = "CardInfoEdit",
    CardStatusChange = "CardStatusChange",
    CardDetailsView = "CardDetailsView",
    CardTopEdit = "CardTopEdit",
    CardBulkOrderCreate = "CardBulkOrderCreate",
    UserOrdersPageView = "UserOrdersPageView",
    UserOrdersStatusChange = "UserOrdersStatusChange",
    UserOrdersDetailsView = "UserOrdersDetailsView",
    BulkOrdersPageView = "BulkOrdersPageView",
    BulkOrdersStatusChange = "BulkOrdersStatusChange",
    BulkOrdersDetailsView = "BulkOrdersDetailsView",
    PartnersOrdersPageView = "PartnersOrdersPageView",
    PartnersOrdersStatusChange = "PartnersOrdersStatusChange",
    PartnersOrdersDetailsView = "PartnersOrdersDetailsView",
    PartnersOrdersOrderCreate = "PartnersOrdersOrderCreate",
    TransactionsPageView = "TransactionsPageView",
    TransactionsStatusChange = "TransactionsStatusChange",
    TransactionsCreateUser = "TransactionsCreateUser",
    TransactionsCreateBulkSell = "TransactionsCreateBulkSell",
    TransactionsCreatePartners = "TransactionsCreatePartners",
    UserVendorsPageView = "UserVendorsPageView",
    UserVendorsDetails = "UserVendorsDetails",
    UserVendorsMoreDetails = "UserVendorsMoreDetails",
    UserVendorsEdit = "UserVendorsEdit",
    UserVendorsBlocking = "UserVendorsBlocking",
    BulkSalesVendorsPageView = "BulkSalesVendorsPageView",
    BulkSalesVendorsDetails = "BulkSalesVendorsDetails",
    BulkSalesVendorsEdit = "BulkSalesVendorsEdit",
    BulkSalesVendorsCreate = "BulkSalesVendorsCreate",
    BulkSalesVendorsBlocking = "BulkSalesVendorsBlocking",
    PartnersSalesVendorsPageView = "PartnersSalesVendorsPageView",
    PartnersSalesVendorsDetails = "PartnersSalesVendorsDetails",
    PartnersSalesVendorsEdit = "PartnersSalesVendorsEdit",
    PartnersSalesVendorsCreate = "PartnersSalesVendorsCreate",
    PartnersSalesVendorsBlocking = "PartnersSalesVendorsBlocking",
    BrandPageView = "BrandPageView",
    BrandTopEdit = "BrandTopEdit",
    BrandEdit = "BrandEdit",
    BrandBlocking = "BrandBlocking",
    BrandCreate = "BrandCreate",
    TaskPageView = "TaskPageView",
    TaskOpenAll = "TaskOpenAll",
    TaskCreate = "TaskCreate",
    TaskCustomerRole = "TaskCustomerRole",
    TaskPartnersRole = "TaskPartnersRole",
    TaskCheckRole = "TaskCheckRole",
    TaskFinanceRole = "TaskFinanceRole",
    SettingsViewCards = "SettingsViewCards",
    SettingsViewDiscount = "SettingsViewDiscount",
    SettingsViewBonusAndFunds = "SettingsViewBonusAndFunds",
    SettingsViewContent = "SettingsViewContent",
    SettingsViewPayment = "SettingsViewPayment",
    AnalyticsViewOverview = "AnalyticsViewOverview",
    AnalyticsViewCards = "AnalyticsViewCards",
    AnalyticsViewBrands = "AnalyticsViewBrands",
    AnalyticsViewValue = "AnalyticsViewValue",
    AnalyticsViewPartners = "AnalyticsViewPartners",
    AuditView = "AuditView"
}